// angular/src/app/api/model/clienteJuridicoResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EnderecoResponse } from './enderecoResponse';


/**
 * Dados completos de um cliente pessoa jurídica
 */
export interface ClienteJuridicoResponse { 
    /**
     * ID do cliente
     */
    readonly id?: number;
    /**
     * Tipo do cliente
     */
    readonly tipo?: ClienteJuridicoResponse.TipoEnum;
    /**
     * CNPJ (formato: 00.000.000/0000-00)
     */
    cnpj?: string;
    /**
     * Razão Social
     */
    razaoSocial?: string;
    /**
     * Inscrição Estadual
     */
    inscricaoEstadual?: string;
    /**
     * E-mail do cliente
     */
    email?: string;
    /**
     * Status do cliente
     */
    estaAtivo?: boolean;
    /**
     * Data de criação da empresa
     */
    dataCriacaoEmpresa?: string;
    /**
     * Lista de endereços do cliente
     */
    enderecos?: Array<EnderecoResponse>;
    /**
     * Data de criação
     */
    readonly createdAt?: string;
    /**
     * Data da última atualização
     */
    readonly updatedAt?: string;
}
export namespace ClienteJuridicoResponse {
    export const TipoEnum = {
        Fisica: 'FISICA',
        Juridica: 'JURIDICA'
    } as const;
    export type TipoEnum = typeof TipoEnum[keyof typeof TipoEnum];
}